Optimize installs and updates
Separate functions in a different file and add bats unit tests.
Add option to use github api token (useful in testing, shouldn't
be necessary in normal installs).
Run pre-install as root (everything in there was already
root dependent).
Improve logging and verbose logging, better showing what is
going on.
Use curl instead of wget everywhere.
Fix Ubuntu 22.04 demand for signing packages.
Fix Google Chrome apt source list double entry by using the same
source list file name "google-chrome.list".

+++ install-user-pkgs.sh
@@ -3,6 +3,7 @@
 set -euo pipefail
 
 BASEDIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+. $BASEDIR/_functions.sh
 
 if [ "$EUID" == "0" ]; then
   echo "Please do not run this script as root"
@@ -10,12 +11,20 @@ if [ "$EUID" == "0" ]; then
 fi
 
 ALL_ARGS=$@
+GH_USERNAME_PASSWORD=''
+CURL_OPTION_GH_USERNAME_PASSWORD=''
 UPDATE=false
 SHOW_HELP=false
 VERBOSE=false
 while [[ $# -gt 0 ]]; do
   key="$1"
   case $key in
+    --gh)
+    GH_USERNAME_PASSWORD=$2
+    CURL_OPTION_GH_USERNAME_PASSWORD=" --user $2 "
+    shift
+    shift
+    ;;
     --update|-u)
     UPDATE=true
     shift
@@ -42,6 +51,7 @@ Usage:
   `readlink -f $0` [flags]
 
 Flags:
+      --gh <user:pw>       GitHub username and password
   -u, --update             Will download and install/reinstall even if the tools are already installed
       --verbose            Show verbose output
   -h, --help               help
@@ -55,10 +65,10 @@ if $VERBOSE; then
 fi
 
 # dvm - deno
-if ! hash dvm 2>/dev/null; then
+if ! hash dvm 2>/dev/null && ! [ -e $HOME/bin/dvm ]; then
+  echo -e "\e[34mInstall DVM.\e[0m"
   DVM_TARGZ='dvm_linux_amd64.tar.gz'
-  DVM_DL_URL=`curl -fsSL https://api.github.com/repos/axetroy/dvm/releases | \
-  jq --raw-output '[.[] | select(.prerelease == false)][0].assets[] | select(.name|test("'$DVM_TARGZ'")).browser_download_url'`
+  DVM_DL_URL=`githubReleaseDownloadUrl axetroy/dvm $DVM_TARGZ`
   DVM_TMP_DIR='/tmp/dvm'
   mkdir -p $DVM_TMP_DIR
   DVM_TARGZ_TMP="$DVM_TMP_DIR/$DVM_TARGZ"
@@ -73,8 +83,10 @@ if ! hash dvm 2>/dev/null; then
   $HOME/bin/dvm use `$HOME/bin/dvm ls | tail -n1`
 elif $UPDATE; then
   dvm upgrade
-  dvm install latest
-  dvm use `dvm ls | tail -n1`
+  if ! dvm ls-remote | grep -q 'Latest and currently using'; then
+    dvm install latest
+    dvm use `dvm ls | tail -n1`
+  fi
 fi
 
 # rbenv
@@ -84,10 +96,8 @@ if ! [ -f $BASEDIR/tools/rbenv/shims/ruby ]; then
   $HOME/.rbenv/bin/rbenv install 2.7.6
   $HOME/.rbenv/bin/rbenv install 3.1.2
   $HOME/.rbenv/bin/rbenv global 3.1.2
-else
-  if $VERBOSE; then
-    echo "Not installing Rbenv and generating Ruby, it is already installed."
-  fi
+elif $VERBOSE; then
+  echo "Not installing Rbenv and generating Ruby, it is already installed."
 fi
 
 # rust
@@ -99,17 +109,37 @@ elif $UPDATE; then
 fi
 
 # tfenv
-if ! $HOME/bin/tfenv list &> /dev/null || $UPDATE; then
+if ! $HOME/bin/tfenv list &> /dev/null; then
+  echo -e "\e[34mInstall Tfenv.\e[0m"
   $BASEDIR/tools/tfenv/bin/tfenv install latest:^0.12.
   $BASEDIR/tools/tfenv/bin/tfenv install latest:^0.13.
   $BASEDIR/tools/tfenv/bin/tfenv install latest
   $BASEDIR/tools/tfenv/bin/tfenv use latest
+elif $UPDATE; then
+  LATEST_TFENV=`getLatestVersion $($BASEDIR/tools/tfenv/bin/tfenv list-remote | grep --color=never -v '-')`
+  LATEST_012=`getLatestVersion $($BASEDIR/tools/tfenv/bin/tfenv list-remote | grep --color=never -v '-' | grep --color=never '^0.12')`
+  LATEST_013=`getLatestVersion $($BASEDIR/tools/tfenv/bin/tfenv list-remote | grep --color=never -v '-' | grep --color=never '^0.13')`
+  CURRENT_012=`$BASEDIR/tools/tfenv/bin/tfenv list | sed -E 's/\*//' | awk '{print $1}' | grep --color=never '^0.12'`
+  CURRENT_013=`$BASEDIR/tools/tfenv/bin/tfenv list | sed -E 's/\*//' | awk '{print $1}' | grep --color=never '^0.13'`
+  echo -e "\e[34mUpdate Tfenv.\e[0m"
+  if [ "$LATEST_012" != "$CURRENT_012" ]; then
+    $BASEDIR/tools/tfenv/bin/tfenv uninstall $CURRENT_012
+    $BASEDIR/tools/tfenv/bin/tfenv install latest:^0.12.
+  fi
+  if [ "$LATEST_013" != "$CURRENT_013" ]; then
+    $BASEDIR/tools/tfenv/bin/tfenv uninstall $CURRENT_013
+    $BASEDIR/tools/tfenv/bin/tfenv install latest:^0.13.
+  fi
+  if ! $BASEDIR/tools/tfenv/bin/tfenv list | sed -E 's/\*//' | awk '{print $1}' | grep --color=never -q $LATEST_TFENV; then
+    $BASEDIR/tools/tfenv/bin/tfenv install latest
+    $BASEDIR/tools/tfenv/bin/tfenv use latest
+  fi
 fi
 
 # krew
-if ! hash krew 2>/dev/null; then
+if ! hash kubectl-krew 2>/dev/null && ! [ -e $HOME/.krew/bin/kubectl-krew ]; then
   echo -e "\e[34mInstall krew.\e[0m"
-  wget https://github.com/kubernetes-sigs/krew/releases/latest/download/krew-linux_amd64.tar.gz -O /tmp/krew.tar.gz
+  curl -fsSL --output /tmp/krew.tar.gz https://github.com/kubernetes-sigs/krew/releases/latest/download/krew-linux_amd64.tar.gz
   rm -rf /tmp/krew/
   mkdir /tmp/krew/
   tar -xvzf /tmp/krew.tar.gz -C /tmp/krew/
@@ -117,50 +147,61 @@ if ! hash krew 2>/dev/null; then
   rm -rf /tmp/krew/
   rm /tmp/krew.tar.gz
 elif $UPDATE; then
+  VERSION=`githubLatestReleaseVersion kubernetes-sigs/krew`
+  CURRENT_VERSION=`kubectl krew version | grep GitTag | awk '{print $2}'`
   kubectl krew upgrade
-else
-  if $VERBOSE; then
-    echo "Not installing Krew, it is already installed."
-  fi
+elif $VERBOSE; then
+  echo "Not installing Krew, it is already installed."
 fi
 
 # docker-show-context
-if ! hash docker-show-context 2>/dev/null || $UPDATE; then
+if ! hash docker-show-context 2>/dev/null && ! [ -e $HOME/bin/docker-show-context ]; then
   echo -e "\e[34mInstall docker-show-context.\e[0m"
   curl -fsSL --output $HOME/bin/docker-show-context https://github.com/pwaller/docker-show-context/releases/latest/download/docker-show-context_linux_amd64
   chmod +x $HOME/bin/docker-show-context
-else
-  if $VERBOSE; then
-    echo "Not installing docker-show-context, it is already installed."
-  fi
+elif $VERBOSE; then
+  echo "Not installing docker-show-context, it is already installed."
 fi
 
 # golang
-if ! hash go &> /dev/null; then
-  GO_ARCH=''
-  case `uname -m` in
-    x86_64)
-      GO_ARCH=amd64
-      ;;
-    aarch64)
-      GO_ARCH=arm64
-      ;;
-    armv7l)
-      GO_ARCH=armv6l
-      ;;
-    *)
-      echo "Golang will not be installed: unsupported architecture: `uname -m`"
-      ;;
-  esac
-  if [ "$GO_ARCH" != '' ]; then
-    GO_VERSION=`curl -fsSL https://api.github.com/repos/golang/go/git/matching-refs/tags/go1. | \
-    jq --raw-output '.[-1].ref' | \
-    sed 's/refs\/tags\/go//'`
-    curl -fsSL --output /tmp/go.tar.gz https://go.dev/dl/go$GO_VERSION.linux-$GO_ARCH.tar.gz
+GO_ARCH=''
+case `uname -m` in
+  x86_64)
+    GO_ARCH=amd64
+    ;;
+  aarch64)
+    GO_ARCH=arm64
+    ;;
+  armv7l)
+    GO_ARCH=armv6l
+    ;;
+  *)
+    echo "Golang will not be installed: unsupported architecture: `uname -m`"
+    ;;
+esac
+if [ "$GO_ARCH" != '' ]; then
+  installGolang () {
+    curl -fsSL --output /tmp/go.tar.gz https://go.dev/dl/go$GO_AVAILABLE_VERSION.linux-$GO_ARCH.tar.gz
     rm -rf $HOME/.go/
     tar -C /tmp/ -xzvf /tmp/go.tar.gz go/bin go/pkg go/lib go/src
     mv /tmp/go $HOME/.go
     rm /tmp/go.tar.gz
+  }
+  GO_AVAILABLE_VERSION=`githubLatestTagByDate golang/go go1. | \
+  sed 's/refs\/tags\/go//'`
+  if ! hash go  2>/dev/null && ! [ -d $HOME/.go/ ] &> /dev/null; then
+    echo -e "\e[34mInstall golang.\e[0m"
+    installGolang
+  elif $UPDATE; then
+    GO_CURRENT_VERSION=`go version 2>/dev/null | sed -E 's/^.*go([[:digit:]]+\.[[:digit:]]+\.[[:digit:]]+).*$/\1/'`
+    if [ "$GO_AVAILABLE_VERSION" != "$GO_CURRENT_VERSION" ]; then
+      echo -e "\e[34mUpdate golang.\e[0m"
+      installGolang
+    elif $VERBOSE; then
+      echo "Not updating golang, it is already up to date."
+    fi
+  elif $VERBOSE; then
+    echo "Not installing golang, it is already installed."
   fi
 fi
 
@@ -185,12 +226,10 @@ case `uname -m` in
     echo "Docker-slim will not be installed: unsupported architecture: `uname -m`"
     ;;
 esac
-if ! hash docker-slim 2>/dev/null; then
+if ! hash docker-slim 2>/dev/null && ! [ -e $HOME/bin/docker-slim ]; then
   echo -e "\e[34mInstall docker-slim.\e[0m"
   DSLIMTAR=/tmp/docker-slim.tar.gz
-  DS_VERSION=`curl -fsSL https://api.github.com/repos/docker-slim/docker-slim/git/matching-refs/tags/1. | \
-  jq --raw-output '.[-1].ref' | \
-  sed 's/refs\/tags\///'`
+  DS_VERSION=`githubLatestTagByVersion docker-slim/docker-slim`
   curl -fsSL --output $DSLIMTAR https://downloads.dockerslim.com/releases/$DS_VERSION/dist_$ARCH.tar.gz
   mkdir /tmp/dslim/
   tar -xvzf $DSLIMTAR -C /tmp/dslim/
@@ -198,9 +237,14 @@ if ! hash docker-slim 2>/dev/null; then
   rm -rf /tmp/dslim/
   rm $DSLIMTAR
 elif $UPDATE; then
-  docker-slim update
-else
-  if $VERBOSE; then
-    echo "Not installing docker-slim, it is already installed."
+  DS_CURRENT_VERSION=`docker-slim --version | cut -d'|' -f3`
+  DS_AVAILABLE_VERSION=`githubLatestTagByVersion docker-slim/docker-slim`
+  if [ "$DS_AVAILABLE_VERSION" != "$DS_CURRENT_VERSION" ]; then
+    echo -e "\e[34mUpdate docker-slim.\e[0m"
+    docker-slim update
+  elif $VERBOSE; then
+    echo "Not updating docker-slim, already up to date."
   fi
+elif $VERBOSE; then
+  echo "Not installing docker-slim, it is already installed."
 fi
